-structuring the code -creating entity for favorite and cart
