--- composeApp/src/androidMain/kotlin/com/virtualworld/mipymeanabel/data/source/local/RoomDataSourceImpl.kt
@@ -1,10 +1,0 @@
-package com.virtualworld.mipymeanabel.data.source.local
-
-import android.content.Context
-
-actual class RoomDataSourceImpl () {
-
-
-
-
-}
