+++ composeApp/src/commonMain/kotlin/com/virtualworld/mipymeanabel/App.kt
@@ -5,7 +5,6 @@ import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.darkColorScheme
 import androidx.compose.material3.lightColorScheme
 import androidx.compose.runtime.Composable
-import com.virtualworld.mipymeanabel.data.source.local.TodoDao
 import com.virtualworld.mipymeanabel.ui.screen.main.MainScreen
 import org.jetbrains.compose.ui.tooling.preview.Preview
