+++ composeApp/src/commonMain/kotlin/com/virtualworld/mipymeanabel/data/dto/ProductInfo.kt
@@ -1,0 +1,12 @@
+package com.virtualworld.mipymeanabel.data.dto
+
+import androidx.room.Entity
+import androidx.room.PrimaryKey
+
+@Entity
+data class ProductInfo(
+    @PrimaryKey(autoGenerate = false)
+    val id: Long,
+    val favorite: Boolean,
+    val cart: Boolean,
+)
