+++ composeApp/src/commonMain/kotlin/com/virtualworld/mipymeanabel/data/repository/remote/FirebaseRepository.kt
@@ -1,7 +1,7 @@
 package com.virtualworld.mipymeanabel.data.repository.remote
 
 import com.virtualworld.mipymeanabel.data.NetworkResponseState
-import com.virtualworld.mipymeanabel.data.model.Product
+import com.virtualworld.mipymeanabel.data.dto.Product
 import kotlinx.coroutines.flow.Flow
 
 interface FirebaseRepository {
