+++ composeApp/src/commonMain/kotlin/com/virtualworld/mipymeanabel/data/repository/remote/FirebaseRepositoryImp.kt
@@ -1,7 +1,7 @@
 package com.virtualworld.mipymeanabel.data.repository.remote
 
 import com.virtualworld.mipymeanabel.data.NetworkResponseState
-import com.virtualworld.mipymeanabel.data.model.Product
+import com.virtualworld.mipymeanabel.data.dto.Product
 import com.virtualworld.mipymeanabel.data.source.remote.FirebaseDataSource
 import kotlinx.coroutines.flow.Flow
