+++ composeApp/src/commonMain/kotlin/com/virtualworld/mipymeanabel/data/source/local/DatabaseMain.kt
@@ -1,64 +1,24 @@
 package com.virtualworld.mipymeanabel.data.source.local
 
 import androidx.room.ConstructedBy
-import androidx.room.Dao
 import androidx.room.Database
-import androidx.room.Entity
-import androidx.room.Insert
-import androidx.room.PrimaryKey
-import androidx.room.Query
 import androidx.room.RoomDatabase
 import androidx.room.RoomDatabaseConstructor
-import androidx.sqlite.driver.bundled.BundledSQLiteDriver
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.IO
-import kotlinx.coroutines.flow.Flow
+import com.virtualworld.mipymeanabel.data.databese.TodoDao
+import com.virtualworld.mipymeanabel.data.dto.ProductInfo
 
-@Database(entities = [TodoEntity::class], version = 1)
+@Database(entities = [ProductInfo::class], version = 1)
 @ConstructedBy(AppDatabaseConstructor::class)
 abstract class AppDatabase : RoomDatabase() {
     abstract fun getDao(): TodoDao
 }
 
-// The Room compiler generates the `actual` implementations.
+
 @Suppress("NO_ACTUAL_FOR_EXPECT")
 expect object AppDatabaseConstructor : RoomDatabaseConstructor<AppDatabase> {
     override fun initialize(): AppDatabase
 }
 
-@Dao
-interface TodoDao {
-    @Insert
-    suspend fun insert(item: TodoEntity)
-
-    @Query("SELECT count(*) FROM TodoEntity")
-    suspend fun count(): Int
-
-    @Query("SELECT * FROM TodoEntity")
-    fun getAllAsFlow(): Flow<List<TodoEntity>>
-}
-
-@Entity
-data class TodoEntity(
-    @PrimaryKey(autoGenerate = true) val id: Long = 0,
-    val title: String,
-    val content: String
-)
 
 
-//expect class DatabaseFactory {
-//    fun createRoomDatabase(): AppDatabase
-//}
 
-// fun getRoomDatabase(
-//    builder: RoomDatabase.Builder<AppDatabase>
-//): AppDatabase {
-//    return builder
-//        .addMigrations()
-//        .fallbackToDestructiveMigrationOnDowngrade(
-//            dropAllTables = TODO()
-//        )
-//        .setDriver(BundledSQLiteDriver())
-//        .setQueryCoroutineContext(Dispatchers.IO)
-//        .build()
-//}
