--- composeApp/src/commonMain/kotlin/com/virtualworld/mipymeanabel/data/source/local/RoomDataSourceImpl.kt
@@ -1,7 +1,0 @@
-package com.virtualworld.mipymeanabel.data.source.local
-
-
-
-expect class RoomDataSourceImpl {
-
-}
