+++ composeApp/src/commonMain/kotlin/com/virtualworld/mipymeanabel/id/NativeModule.kt
@@ -1,0 +1,5 @@
+package com.virtualworld.mipymeanabel.id
+
+import org.koin.core.module.Module
+
+expect val nativeModule: Module
