+++ composeApp/src/commonMain/kotlin/com/virtualworld/mipymeanabel/id/StartModule.kt
@@ -1,0 +1,11 @@
+package com.virtualworld.mipymeanabel.id
+
+import org.koin.core.context.startKoin
+import org.koin.dsl.KoinAppDeclaration
+
+fun  initKoin(configuration: KoinAppDeclaration? = null) {
+    startKoin{
+        configuration?.invoke(this)
+        modules(viewModelsModule,dataModules, nativeModule)
+    }
+}
