+++ composeApp/src/iosMain/kotlin/com/virtualworld/mipymeanabel/data/source/local/Database.kt
@@ -1,27 +1,13 @@
 package com.virtualworld.mipymeanabel.data.source.local
 
-import androidx.room.RoomDatabase
 import androidx.room.Room
+import androidx.room.RoomDatabase
 import androidx.sqlite.driver.bundled.BundledSQLiteDriver
 import kotlinx.cinterop.ExperimentalForeignApi
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.IO
 import platform.Foundation.NSDocumentDirectory
 import platform.Foundation.NSFileManager
 import platform.Foundation.NSUserDomainMask
 
-
-//actual class DatabaseFactory {
-//    actual fun createRoomDatabase(): AppDatabase {
-//        val dbFile = documentDirectory() + "/my_room.db"
-//        return Room.databaseBuilder<AppDatabase>(
-//            name = dbFile,
-//        )
-//            .setDriver(BundledSQLiteDriver())
-//            .setQueryCoroutineContext(Dispatchers.IO)
-//            .build()
-//    }
-//}
 
 fun getDatabaseBuilder(): RoomDatabase.Builder<AppDatabase> {
     val dbFilePath = documentDirectory() + "/my_room.db"
