--- composeApp/src/iosMain/kotlin/com/virtualworld/mipymeanabel/data/source/local/RoomDataSourceImpl.ios.kt
@@ -1,5 +1,0 @@
-package com.virtualworld.mipymeanabel.data.source.local
-
-actual class RoomDataSourceImpl {
-
-}
